\name{init}
\alias{init}
\title{Initialise Wiki Workflow Directory}
\usage{
init(Makefile = "Makefile", clone = TRUE)
}
\arguments{
  \item{Makefile}{Filename to add rules to}

  \item{clone}{Try to clone the repository immediately on
  initialising?}
}
\description{
Initialise a directory to work with my wiki workflow
approach
}
\author{
Rich FitzJohn
}

